http_traffic->http rename.
